Fix CID 164507,8 (Unintentional integer overflow)

+++ src/grdfill.c
@@ -338,8 +338,8 @@ int GMT_grdfill (void *V_API, int mode, void *args) {
 	
 	ID = gmt_M_memory_aligned (GMT, NULL, Grid->header->size, char);
 	/* Set the top and bottom boundary rows to UINT_MAX */
-	offset = (Grid->header->pad[YHI] + Grid->header->n_rows) * Grid->header->mx;
-	for (node = 0; node < Grid->header->pad[YHI]*Grid->header->mx; node++) ID[node] = ID[node+offset] = 1;
+	offset = (uint64_t)(Grid->header->pad[YHI] + Grid->header->n_rows) * Grid->header->mx;
+	for (node = 0; node < (uint64_t)Grid->header->pad[YHI]*Grid->header->mx; node++) ID[node] = ID[node+offset] = 1;
 	/* Set the left and right boundary columnss to UINT_MAX */
 	offset = Grid->header->pad[XLO] + Grid->header->n_columns;
 	for (row = 0; row < Grid->header->my; row++) {
